chore: add `get_bytecode`, `get_calldata` to args
