+++ crates/cfg/src/core/mod.rs
@@ -2,10 +2,7 @@ pub(crate) mod graph;
 
 use alloy::primitives::Address;
 use eyre::eyre;
-use heimdall_common::{
-    ether::{bytecode::get_bytecode_from_target, compiler::detect_compiler},
-    utils::strings::StringExt,
-};
+use heimdall_common::{ether::compiler::detect_compiler, utils::strings::StringExt};
 use heimdall_vm::core::vm::VM;
 
 use petgraph::{dot::Dot, Graph};
@@ -53,7 +50,8 @@ pub async fn cfg(args: CFGArgs) -> Result<CFGResult, Error> {
 
     // get the bytecode from the target
     let start_fetch_time = Instant::now();
-    let contract_bytecode = get_bytecode_from_target(&args.target, &args.rpc_url)
+    let contract_bytecode = args
+        .get_bytecode()
         .await
         .map_err(|e| Error::FetchError(format!("fetching target bytecode failed: {}", e)))?;
     debug!("fetching target bytecode took {:?}", start_fetch_time.elapsed());
