+++ crates/cli/Cargo.toml
@@ -27,6 +27,7 @@ tracing-subscriber = "0.3.18"
 eyre = "0.6.12"
 alloy-json-abi = "0.7.6"
 alloy = { version = "0.1.3", features = ["full", "rpc-types-debug", "rpc-types-trace"] }
+async-trait = "0.1.51"
 
 [[bin]]
 name = "heimdall"
